Rename "Adapter" to "Authenticator"
