+++ src/Authenticator/FormAuthenticator.php
@@ -11,7 +11,7 @@
  * @since         4.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Authentication\Adapter;
+namespace Authentication\Authenticator;
 
 use Authentication\Result;
 use Psr\Http\Message\ResponseInterface;
