+++ tests/TestCase/Authenticator/HttpBasicAuthenticatorTest.php
@@ -11,9 +11,9 @@
  * @since         4.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Authentication\Test\TestCase\Adapter;
+namespace Authentication\Test\TestCase\Authenticator;
 
-use Authentication\Adapter\HttpBasicAuthenticator;
+use Authentication\Authenticator\HttpBasicAuthenticator;
 use Authentication\Identifier\IdentifierCollection;
 use Authentication\Test\TestCase\AuthenticationTestCase as TestCase;
 use Cake\Http\ServerRequestFactory;
